VFB-43: Refactor the Add Parcel Form to be shared ParcelForm
